Enemy Code partially complete
Complete the code for the enemy. Now, enemy have a sight range, and will attack player when he enter that sight. After having the animation for enemy, some part of the enemy code need to be changed to follow the animation.

Also, Health system are added. Already add the reference to the enemy. The health code should also be referenced in Player code and to count for the health.

Health UI need to be complete later.